+++ mpqp/execution/providers/aws.py
@@ -260,7 +260,7 @@ def extract_result(
             nb_qubits = device_params.paradigmParameters.qubitCount
             shots = braket_result.task_metadata.shots
             measure = ExpectationMeasure(
-                Observable(np.zeros((2**nb_qubits, 2**nb_qubits), dtype=np.complex64)),
+                Observable(np.zeros((2**nb_qubits, 2**nb_qubits), dtype=np.complex128)),
                 list(range(nb_qubits)),
                 shots,
             )
