+++ tests/core/instruction/gates/test_gate.py
@@ -69,7 +69,7 @@ def test_is_not_equivalent(g1: Gate, g2: Gate):
         ),
     ],
 )
-def test_power(gate: Gate, pow: float, result_matrix: npt.NDArray[np.complex64]):
+def test_power(gate: Gate, pow: float, result_matrix: npt.NDArray[np.complex128]):
     assert matrix_eq(gate.power(pow).to_matrix(), result_matrix)
 
 
@@ -83,7 +83,7 @@ def test_power(gate: Gate, pow: float, result_matrix: npt.NDArray[np.complex64])
         ),
     ],
 )
-def test_tensor_product(g1: Gate, g2: Gate, result_matrix: npt.NDArray[np.complex64]):
+def test_tensor_product(g1: Gate, g2: Gate, result_matrix: npt.NDArray[np.complex128]):
     with pytest.warns(NumberQubitsWarning):
         assert matrix_eq(g1.tensor_product(g2).to_matrix(), result_matrix)
 
@@ -94,7 +94,7 @@ def test_tensor_product(g1: Gate, g2: Gate, result_matrix: npt.NDArray[np.comple
         (X(0), Z(0), np.array([[0, -1], [1, 0]])),
     ],
 )
-def test_product(g1: Gate, g2: Gate, result_matrix: npt.NDArray[np.complex64]):
+def test_product(g1: Gate, g2: Gate, result_matrix: npt.NDArray[np.complex128]):
     assert matrix_eq(g1.product(g2).to_matrix(), result_matrix)
 
 
@@ -105,7 +105,7 @@ def test_product(g1: Gate, g2: Gate, result_matrix: npt.NDArray[np.complex64]):
     ],
 )
 def test_scalar_product(
-    g1: Gate, scalar: complex, result_matrix: npt.NDArray[np.complex64]
+    g1: Gate, scalar: complex, result_matrix: npt.NDArray[np.complex128]
 ):
     assert matrix_eq(g1.scalar_product(scalar).to_matrix(), result_matrix)
 
@@ -116,7 +116,7 @@ def test_scalar_product(
         (X(0), Z(0), np.array([[1, 1], [1, -1]]) / np.sqrt(2)),
     ],
 )
-def test_add_product(g1: Gate, g2: Gate, result_matrix: npt.NDArray[np.complex64]):
+def test_add_product(g1: Gate, g2: Gate, result_matrix: npt.NDArray[np.complex128]):
     assert matrix_eq(g1.plus(g2).to_matrix(), result_matrix)
 
 
@@ -126,7 +126,7 @@ def test_add_product(g1: Gate, g2: Gate, result_matrix: npt.NDArray[np.complex64
         (X(0), Z(0), np.array([[-1, 1], [1, 1]]) / np.sqrt(2)),
     ],
 )
-def test_sub_product(g1: Gate, g2: Gate, result_matrix: npt.NDArray[np.complex64]):
+def test_sub_product(g1: Gate, g2: Gate, result_matrix: npt.NDArray[np.complex128]):
     assert matrix_eq(g1.minus(g2).to_matrix(), result_matrix)
